+++ src/main/java/roborally/events/AnimateEvent.java
@@ -7,8 +7,8 @@ import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.scenes.scene2d.ui.Image;
 import roborally.game.IGame;
 import roborally.game.robot.Robot;
-import roborally.gameview.elements.ProgramCardsView;
-import roborally.gameview.elements.UIElements;
+import roborally.gameview.ui.ProgramCardsView;
+import roborally.gameview.ui.UIElements;
 import roborally.utilities.SettingsUtil;
 
 import java.util.ArrayList;
