+++ src/main/java/roborally/game/Game.java
@@ -16,8 +16,8 @@ import roborally.game.robot.IRobot;
 import roborally.game.robot.Robot;
 import roborally.game.structure.IRound;
 import roborally.game.structure.Round;
-import roborally.gameview.elements.ProgramCardsView;
-import roborally.gameview.elements.UIElements;
+import roborally.gameview.ui.ProgramCardsView;
+import roborally.gameview.ui.UIElements;
 import roborally.gameview.layout.ILayers;
 import roborally.gameview.layout.Layers;
 import roborally.utilities.AssetManagerUtil;
