+++ src/main/java/roborally/game/IGame.java
@@ -3,7 +3,7 @@ package roborally.game;
 import roborally.game.robot.Robot;
 import roborally.game.structure.IRound;
 import roborally.gameview.layout.ILayers;
-import roborally.gameview.elements.ProgramCardsView;
+import roborally.gameview.ui.ProgramCardsView;
 
 import java.util.ArrayList;
