+++ src/main/java/roborally/gameview/GameView.java
@@ -19,8 +19,8 @@ import roborally.events.AnimateEvent;
 import roborally.events.Events;
 import roborally.game.Game;
 import roborally.game.IGame;
-import roborally.gameview.elements.ProgramCardsView;
-import roborally.gameview.elements.UIElements;
+import roborally.gameview.ui.ProgramCardsView;
+import roborally.gameview.ui.UIElements;
 import roborally.gameview.menu.Menu;
 import roborally.utilities.AssetManagerUtil;
 import roborally.utilities.SettingsUtil;
