+++ src/main/java/roborally/gameview/ui/ProgramCardsView.java
@@ -1,4 +1,4 @@
-package roborally.gameview.elements;
+package roborally.gameview.ui;
 
 import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.g2d.BitmapFont;
