+++ src/main/java/roborally/gameview/ui/elements/IElement.java
@@ -1,4 +1,4 @@
-package roborally.gameview.elements;
+package roborally.gameview.ui.elements;
 
 import com.badlogic.gdx.scenes.scene2d.ui.Image;
