+++ src/main/java/roborally/gameview/ui/elements/Message.java
@@ -1,4 +1,4 @@
-package roborally.gameview.elements;
+package roborally.gameview.ui.elements;
 
 import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.g2d.BitmapFont;
