+++ src/main/java/roborally/gameview/ui/elements/buttons/RestartButton.java
@@ -1,11 +1,11 @@
-package roborally.gameview.elements.buttons;
+package roborally.gameview.ui.elements.buttons;
 
 import com.badlogic.gdx.scenes.scene2d.InputEvent;
 import com.badlogic.gdx.scenes.scene2d.ui.ImageButton;
 import com.badlogic.gdx.scenes.scene2d.utils.ClickListener;
 import com.badlogic.gdx.scenes.scene2d.utils.TextureRegionDrawable;
 import roborally.game.IGame;
-import roborally.gameview.elements.UIElements;
+import roborally.gameview.ui.UIElements;
 import roborally.utilities.SettingsUtil;
 
 import static roborally.utilities.enums.UIElement.RESTART_BUTTON;
